fix(ci): speed up frontend guardrails
